WIP: AHCI: get it to build and almost work.
This is a working copy that has not been cleaned up. Please look at my
next patch for what's current. I have a cleaner version of this commit
on another branch, so this will eventually be dropped.

In qemu, it still shows the device as having zero bytes.
But it does find it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Signed-off-by: Fergus Simpson <afergs@google.com>