WIP: Testing changes trying to get AHCI driver working.
Note that this is just a snapshot of my current working code and I don't
mean for this to be submitted. A cleaner version with better broken down
commits is being worked on in another branch. This code is closer to
working though so it's what you'll want to use for debugging AHCI until
my other branch is caught up and I can abandon this change.