AHCI: Prevent sign extension of partial address
Drive reads were not working past the 1 TiB mark because the resulting
address negative. This was determined to be an issue with an unsigned
char getting sign extended when bit shifted into a int64_t. It is now
cast to a uint32_t after the shift to prevent sign extension. The
container was also changed from int64_t to uint64_t.

Change-Id: I590b0da4fd0c02b0e2542a0b65bde510bba89525
Signed-off-by: Fergus Simpson <afergs@google.com>